+++ src/model/Stripe.cc
@@ -1,6 +1,6 @@
 #include "Stripe.hh"
 
-Stripe::Stripe(int id, vector<int> &stripe_indices)
+Stripe::Stripe(size_t id, vector<size_t> &stripe_indices)
 {
     _id = id;
     _stripe_indices = stripe_indices;
@@ -10,18 +10,18 @@ Stripe::~Stripe()
 {
 }
 
-vector<int> &Stripe::getStripeIndices() {
+vector<size_t> &Stripe::getStripeIndices() {
     return _stripe_indices;
 }
 
-int Stripe::getId() {
+size_t Stripe::getId() {
     return _id;
 
 }
 
 void Stripe::print() {
     for (auto idx : _stripe_indices) {
-        printf("%d ", idx);
+        printf("%ld ", idx);
     }
     printf("\n");
 }
